feat: Add authentication security in the Item controllers

+++ src/controllers/handlers/item/ItemRD.java
@@ -9,6 +9,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.ServletException;
 
 import models.Item;
+import utils.AuthenticationSecurity;
 import utils.ExceptionHandler;
 import utils.RequestChecker;
 
@@ -16,7 +17,7 @@ public class ItemRD extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         try {
-            if (RequestChecker.isDeleteMode(req)) {
+            if (RequestChecker.isDeleteMode(req) && AuthenticationSecurity.isLoggedIn(req)) {
                 int itemId = Integer.parseInt(req.getParameter("item-id"));
                 Item.deleteById(itemId);
             }
@@ -34,7 +35,8 @@ public class ItemRD extends HttpServlet {
         }
     }
 
-    private void setAttributes(HttpServletRequest req, String searchKeyword) throws ClassNotFoundException, SQLException {
+    private void setAttributes(HttpServletRequest req, String searchKeyword)
+            throws ClassNotFoundException, SQLException {
         if (searchKeyword != null && !searchKeyword.isEmpty()) {
             req.setAttribute("item-list", Item.searchItem(searchKeyword));
         } else {
